+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Added support for multiple prelude files to install and release scripts and Makefile.
 * Added `__POS__` meta operation that can be used to report current position in a file.
 * Added 'Division by zero error' to the interpreter and constants for floating point infinity.
+* Added the `global` operation that stashes values in the global scope.
+
 ## 0.2.2 -- 2024-08-11
 
 * **MAJOR BUG FIX!** The append operation will now be parsed as "++" and not two "+" as in version 0.2.1